Button test successful
Successfully connected a button to a function in SampleController.java.
We now know how to assign functions to buttons.
Next we need to learn how a function can read values from textfields and
how it can create and append an object to the scene
(such as adding a label object to the result area)